% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sga.R
\name{checkinputs}
\alias{checkinputs}
\title{Check inputs to function.}
\usage{
checkinputs(data, prob, S, G)
}
\arguments{
\item{data}{Past data realisations as vectors in a list.  Each list element corresponds to a period of training data.}

\item{prob}{List of functions to simulate from probabilistic forecasts.  Each list element corresponds to a period of training data. The output of each function should be a matrix.}

\item{S}{Matrix encoding linear constraints.}

\item{G}{Values of reconciliation parameters \eqn{a} and \eqn{G} where \eqn{\tilde{y}=S(a+G\hat{y})}.  The first \eqn{m} elements correspond to translation vector \eqn{a}, while the remaining elements correspond to the matrix \eqn{G} where the elements are filled in column-major order.}
}
\description{
This function checks that the inputs for \code{\link[ProbReco]{scoreopt}} and \code{\link[ProbReco]{energy_score}} are correctly setup.  It is called at the start of \code{\link[ProbReco]{scoreopt}}.
}
