% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{energy_i}
\alias{energy_i}
\title{Computes energy score contribution with unconstrained G}
\usage{
energy_i(Sin, yin, xin, xsin, Gin)
}
\arguments{
\item{Sin}{Matrix encoding linear constraints}

\item{yin}{Vector valued realisation}

\item{xin}{Vector valued realisation from base predictive}

\item{xsin}{Copy of vector valued realisation from base predictive}

\item{Gin}{Reconciliation matrix (vectorised)}
}
\value{
Contribution to energy score and gradient w.r.t G
}
\description{
Computes energy score contribution with unconstrained G
}
\examples{
S<-matrix(c(1,1,1,0,0,1),3,2, byrow = TRUE)
Gvec<-as.matrix(runif(6))
y<-as.matrix(S\%*\%runif(2))
x<-as.matrix(runif(3))
xs<-as.matrix(runif(3))
energy_i(S,y,x,xs,Gvec)
}
