% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sga.R
\name{energy_score}
\alias{energy_score}
\title{Find estimate of total energy score and its gradient for a reconciled forecast}
\usage{
energy_score(S, Gvec, data, prob, Q = 500)
}
\arguments{
\item{S}{Matrix encoding linear constraints.}

\item{Gvec}{G Matrix for reconciliation vectorised.}

\item{data}{Data realisations as vector in a list.}

\item{prob}{Functions to simulate from probabilistic forecast (in a list).}

\item{Q}{number of draws for each time period}
}
\value{
Contribution to energy score and gradient w.r.t G.
}
\description{
Find estimate of total energy score and its gradient for a reconciled forecast
}
\examples{
library(purrr)
S<-matrix(c(1,1,1,0,0,1),3,2, byrow = TRUE)
Gvec<-as.matrix(runif(6))
data<-map(1:100,function(i){S\%*\%rnorm(2)})
prob<-map(1:100,function(i){f<-function(){rnorm(3)}})
energy_score(S,Gvec,data,prob)
}
