% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sga.R
\name{find_grad}
\alias{find_grad}
\title{Find estimate of gradient for scores}
\usage{
find_grad(S, Gvec, data, prob, Q)
}
\arguments{
\item{S}{Matrix encoding linear constraints.}

\item{Gvec}{G Matrix for reconciliation vectorised.}

\item{data}{Data realisations as vector in a list.}

\item{prob}{Functions to simulate from probabilistic forecast (in a list).}

\item{Q}{number of draws for each time period}
}
\value{
Contribution to energy score and gradient w.r.t G.
}
\description{
Find estimate of gradient for scores
}
\examples{
library(purrr)
S<-matrix(c(1,1,1,0,0,1),3,2, byrow = TRUE)
Gvec<-as.matrix(runif(6))
data<-map(1:10,function(i){S\%*\%rnorm(2)})
prob<-map(1:10,function(i){f<-function(){rnorm(3)}})
find_grad(S,Gvec,data,prob,20)
}
