% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbReco-package.R
\docType{package}
\name{ProbReco-package}
\alias{ProbReco}
\alias{ProbReco-package}
\title{ProbReco: Probabilistic Forecast Reconciliation based on Score Optimisation}
\description{
Finds a linear reconciliation matrix by optimising with respect to a proper scoring rule.  More description to come as features are added.
}
\details{
This package will carry out probabilistic forecast reconciliation via score 
optimisation.  The only function at the moment is  [energy_i()].
}
\keyword{internal}
