% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sga.R
\name{energy_score}
\alias{energy_score}
\title{Find estimate of total energy score and its gradient for a reconciled forecast}
\usage{
energy_score(data, prob, S, Gvec, Q = 500)
}
\arguments{
\item{data}{Past data realisations as vectors in a list.  Each list element corresponds to a period of training data.}

\item{prob}{Functions to simulate from probabilistic forecast in a list.  Each list element corresponds to a period of training data.}

\item{S}{Matrix encoding linear constraints.}

\item{Gvec}{G Matrix for reconciliation vectorised.}

\item{Q}{Number of draws for each time period used to estimate energy score. Default is 500.}
}
\value{
Total energy score and gradient w.r.t G.
\item{grad}{The estimate of the gradient.}
\item{value}{The estimated total energy score.}
}
\description{
Find estimate of total energy score and its gradient for a reconciled forecast
}
\examples{
library(purrr)
S<-matrix(c(1,1,1,0,0,1),3,2, byrow = TRUE)
Gvec<-as.matrix(runif(6))
data<-map(1:100,function(i){S\%*\%rnorm(2)})
prob<-map(1:100,function(i){f<-function(){rnorm(3)}})
energy_score(data,prob,S,Gvec)
}
\seealso{
Other sga: \code{\link{scoreopt.control}},
  \code{\link{scoreopt}}
}
\concept{sga}
