\name{ProbReco-package}
\alias{ProbReco-package}
\alias{ProbReco}
\docType{package}
\title{
\packageTitle{ProbReco}
}
\description{
\packageDescription{ProbReco}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ProbReco}
\packageIndices{ProbReco}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ProbReco}

Maintainer: \packageMaintainer{ProbReco}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
