% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbReco-package.R
\docType{package}
\name{ProbReco-package}
\alias{ProbReco}
\alias{ProbReco-package}
\title{ProbReco: Score Optimal Probabilistic Forecast Reconciliation}
\description{
Training of reconciliation weights for probabilistic forecasts to optimise energy score using Stochastic Gradient Ascent with automatically differentiated gradients.
}
\details{
This package carries out probabilistic forecast reconciliation via score 
optimisation.  Given incoherent (base) probabilistic forecasts formed over a training 
data set, the function \code{\link[ProbReco]{scoreopt}} finds linear reconciliation weights that optimise 
total energy score over the training data.  The optimisation is carried out using 
the Adaptive Moments (Adam) variant of Stochastic Gradient Ascent.  Tuning parameters
for this optimisation can be set using \code{\link[ProbReco]{scoreopt.control}}.  The gradients are found 
using the automatic differentiation libraries of the Stan project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anastasiospanagiotelis/ProbReco}
  \item Report bugs at \url{https://github.com/anastasiospanagiotelis/ProbReco/issues}
}

}
\keyword{internal}
