% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sga.R
\name{scoreopt.control}
\alias{scoreopt.control}
\title{Tuning parameters for score optimisation by Stochastic Gradient Ascent}
\usage{
scoreopt.control(Q = 500, alpha = 0.1, beta1 = 0.9, beta2 = 0.999,
  maxIter = 500, tol = 0.1, epsilon = 1e-08)
}
\arguments{
\item{Q}{Number of draws for each time period used to estimate energy score. Default is 500.}

\item{alpha}{Learning rate. Deafult is 0.1}

\item{beta1}{Forgetting rate for mean. Default is 0.9.}

\item{beta2}{Forgetting rate for variance. Default is 0.999.}

\item{maxIter}{Maximum number of iterations. Default is 500}

\item{tol}{Tolerance for stopping criterion. Algorithm stops when the change in all parameter values is less than this amount. Default is 0.1.}

\item{epsilon}{small constant added to denominator of step size. Default is 1e-8}
}
\description{
Function to set tuning parameters for stochastic gradient ascent used to
find a reconciliation matrix that optimised a scoring function.
}
\examples{
#Change Maximum Iterations to 1000
scoreopt.control(maxIter=1000)
}
\seealso{
Other ProbReco functions: \code{\link{energy_score}},
  \code{\link{scoreopt}}
}
\concept{ProbReco functions}
